requirements.txt
# MCP Server Dependencies
fastmcp>=0.1.0
requests>=2.31.0
ics>=0.7.2

# Optional: for enhanced functionality
google-generativeai>=0.3.0  # For Google AI Studio integration
python-dotenv>=1.0.0        # For environment variables
uvicorn>=0.24.0             # ASGI server for development
